refactor: move appointments files to scheluder directory
