+++ src/app.module.ts
@@ -4,8 +4,8 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { DatabaseModule } from './database/database.module';
 import { SchedulerModule } from './scheduler/scheduler.module';
-import { AppointmentsController } from './appointments/appointments.controller';
-import { AppointmentsService } from './appointments/appointments.service';
+import { AppointmentsController } from './scheduler/appointments.controller';
+import { AppointmentsService } from './scheduler/appointments.service';
 
 @Module({
   imports: [
