+++ src/scheduler/scheduler.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AppointmentsController } from '../appointments/appointments.controller';
-import { AppointmentsService } from '../appointments/appointments.service';
+import { AppointmentsController } from './appointments.controller';
+import { AppointmentsService } from './appointments.service';
 
 @Module({
   controllers: [AppointmentsController],
